+++ atlasdb-coordination-impl/src/main/java/com/palantir/atlasdb/coordination/CoordinationServiceImpl.java
@@ -25,6 +25,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.common.collect.Iterables;
+import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.keyvalue.impl.CheckAndSetResult;
 import com.palantir.logsafe.SafeArg;
 
@@ -48,6 +49,11 @@ public class CoordinationServiceImpl<T> implements CoordinationService<T> {
         return Optional.of(cachedReference);
     }
 
+    /**
+     * In case of failure, the returned value and bound are guaranteed to be the ones that were in the KVS at the time
+     * of CAS failure only if {@link KeyValueService#getCheckAndSetCompatibility()} is
+     * {@link com.palantir.atlasdb.keyvalue.api.CheckAndSetCompatibility#SUPPORTED_DETAIL_ON_FAILURE}.
+     */
     @Override
     public CheckAndSetResult<ValueAndBound<T>> tryTransformCurrentValue(Function<ValueAndBound<T>, T> transform) {
         CheckAndSetResult<ValueAndBound<T>> transformResult = store.transformAgreedValue(transform);